Rename a file from js to ts
